refactor(core)!: one score scale for a search candidate and an identity (#319)
* refactor(core)!: one score scale for a search candidate and an identity

SearchCandidate.score was MusicBrainz's own 0-100 relevance leaking onto the
published surface, which every other score on it gives as a Float on 0.0-1.0.
It becomes matchScore: Float, normalised and clamped at the mapper, and the
constructor is reordered required-first. IdentityResolution.matchScore and the
three profiles' identityMatchScore follow it onto the same scale.

* docs: name matchScore where the shipped surface still said score

+++ musicmeta-core/src/main/kotlin/com/landofoz/musicmeta/ArtistProfile.kt
@@ -13,7 +13,7 @@ public data class ArtistProfile(val name: String, val results: EnrichmentResults
     // --- Identity ---
     val identifiers: EnrichmentIdentifiers get() = results.identity.identifiers
     val canonicalStatus: CanonicalStatus get() = results.identity.status
-    val identityMatchScore: Int? get() = results.identity.matchScore
+    val identityMatchScore: Float? get() = results.identity.matchScore
     val suggestions: List<SearchCandidate> get() = results.identity.suggestions
 
     // --- Artwork ---
